Add Tags type and NoFinalize() capability to ID, Tag and Tags (#164)
Tags is a new type that encapsulates a []Tag slice and a pool so it can be passed around and pooled without having to hold a reference to the pool. One can simply Finalize() the Tags type and have the slice be returned to the pool.

NoFinalize() is to be used when longly-lived types are being passed to sub-systems that have a mix of longly-lived types and shortly lived types and would like to uniformly finalize them all. Components can return their types marked with NoFinalize() so that calling Finalize() on them is a no-op and can be safely returned without worry that the caller will impact the lifecycle of the type.