Add retry options for creating etcd client (#232)
When M3 processes using m3cluster for service discovery comes up, a lot of processes will log a fatal error message and fail if they cannot start up.  

This is fine but can cause a large amount of DNS requests to be issued if restarted by a processes supervisor very frequently.  To avoid some of these thundering herd issues if a KV client is down, we add some level of backoff when trying to connect to service discovery.