[docs] Fix docs deploy target (#835)
The `docs-deploy` target was broken because 1) the version of default
version of git in alpine linux is lightweight and doesn't include
`fast-import`, which `mkdocs` relies on and 2) even if that was the case
the git process in the container couldn't push to github due to lack of
ssh key access. This fixes those issues.