[coordinator] placement: adds "safe" by default
This PR modifies the `placement/add` handler such that adds are "safe"
by default. That is, we know that operationally it's unsafe to add an
instance to the placement while others are bootstrapping. This encodes
that in the add handler by ensuring that adds will only succeed if there
are no instances initializing, unless the user has specified `force` in
the request.