[coordinator] placement: add+remove fully atomic
I realized that even with the change to make adds "safe" by default,
there could still be a race. That is, the placement that all shards were
validated as AVAILABLE for could not be the present placement by the
time `AddInstances` was called.

This changes placement adds/removes to use the placement algorithm to
construct the new placement, and perform a `CheckAndSet` with the old
placement.

Additionally, this implements "safe" adds in the delete handler as well.