Use a default pool size of 10 in development for query (#1119)
I found that when I got panics running locally, the stack trace was huge due to the large number of goroutines we had spawned. Most of them were from the readWorkerPool and writeWorkerPool in query, since default config gives you 4096 threads. I added config to the local query config to shrink this down to a starting size of 10, which is much more manageable.