[hostid] support reading ID from a file (#204)
Under some orchestration systems, such as Kubernetes, we may need to
read host identity from a file on disk that is dynamically populated
with instance metadata. Additionally, this metadata may sometimes be
populated by an external actor and may not be available at startup, so
we provide an optional timeout to wait for it.

As a concrete example, on Kubernetes this supports reading host ID from
pod metadata that is injected using [downward API
volumes](https://git.io/fp4hL). The host ID may not be immediately
populated, but the user can provide a timeout that the instance will
wait for it before failing to start. An example of a pod metadata file
in Kubernetes is:

```
root@matt-test-656988ff69-gk49b:/# cat /etc/podinfo/id
test-id
```