Port the x/cost package from statsdex (#1215)
This is a prerequisite diff for cost accounting in m3query. It lifts the cost package from statsdex, modifying it slightly to work in the m3 monorepo context, and tweaking some interfaces that I found inconvenient. I use this in #1207 to implement datapoint based cost accounting.

Note: I haven't maintained the statsdex commit history here, but @jeromefroe is the actual author of most of this code.