Enable and fix carbon and graphite docker integration test (#1319)
This also updates the carbon ingester configuration to match the default M3 ingester config where if the config is not nil, then it is used, otherwise it's not. This removes the enabled field which was rather different to how the M3 ingester config's presence or absence defines whether to use it or not.