Copy cors from etcd, since they eliminated the package (#1320)
A small hack I did caught up to me--etcd had a reasonable cors middleware implementation, which I added to m3query. They later removed the package (not surprisingly--it was a random thing to expose). This causes glide up to fail. To fix this, I copied in the package to x/net/http. We could instead add a different dependency (e.g. https://github.com/rs/cors). I mostly went with the copy route because the package is quite small and it didn't seem worth the added dep.