[docker] add image for setting recommended sysctls
This PR provides users with a docker image and Kubernetes DaemonSet that
will set the host's sysctls to our recommended settings. We try to be
careful about this and recommend pinning to a docker image SHA so that
the content is immutable (i.e. someone who controlled our Quay repo
couldn't run arbitrary root containers on hosts that have the image).