Cost accounting hooked up with removal of cost after query finishes (#1207)
This PR tracks and limits datapoint usage in the query engine. Tracking/limiting occurs at both the per-query and global levels. This is the second of 2 PR's, building on top of @jeromefroe 's x/cost package which I lifted and modified from statsdex.

It adds a query/cost package to track both per query and global resource usage, and integrates it through the query lifecycle (starting at engine, and going down to accounted_series_iter.go).

The approach here is to hook a cost.ChainedEnforcer instance into creation of any blocks. Cost is added when datapoints are added to a block; cost is reduced when the block is closed (ChainedEnforcer.Release()). Datapoints are added:

- At the decompression step (see accounted_series_iter.go), since that's the first point at which we know datapoints have actually been used.
- In ColumnBlockBuilder, since we allocate further slices of datapoints at that point.

The datapoints are considered "in use" until the block is closed.

Currently the code reports these stats:

```
cost.global.datapoints: gauge; the number of datapoints currently in use
//                           by this instance.
//
//   cost.global.datapoints_counter: counter; counter representation of the
//                                   number of datapoints in use by this instance
//
//   cost.{per_query,global}.over_datapoints_limit: counter; how many queries are over the
//    datapoint limit
//
//   cost.per_query.max_datapoints_hist: histogram; represents the
//     distribution of the maximum datapoints used at any point in each query.
```