[coordinator] disable stack traces in logs
It's often clear where errors are coming from. When stack traces are
enabled, if a coordinator is being used as a prom backend and there's an
error with the M3DB cluster then every single prom write causes a stack
trace and the logs quickly fill up, making debugging difficult.