[ci] docs: also build operator docs
We serve the operator's docs from a branch subdomain of the monorepo.
Let's incorporate it into the build process.

This means that once we update docs in the operator repo (which will
eventually be merged w/ the monorepo anyway) we'll have to wait until
the next master build, but that slight drift is fine.