Emit single Gauge for bootstrapper order, add tests
Bootstrapper order is emitted in a single Gauge like:

```
bootstrap_bootstrappers{bootstrapper1="filesystem", bootstrapper2="peers",...} <float64>
```

On startup, the labels are set to `cfg.Bootstrappers` and the value of the metric is `1`.

In case of an update, a new metric with the same name but different labels gets emitted
(value `1`) and the value of the original Gauge is set to `0`.