[docker] builder image base alpine 3.9
We have to use an older version of alpine, as git 2.22 removes the '-x'
flag to a submodule command which in turn breaks glide, the old AF
dependency tool we have. Until we migrate to a new version of glide or
go modules, we need to use this older base image.