[aggregator] Make YAML serialization roundtrip for config types (#1864)
What this PR does / why we need it: Many of our types (24 in my last count) implement UnmarshalYAML but not MarshalYAML. This means (in most cases) that the types don't "roundtrip" in YAML--you can't serialize to YAML and get the same object back.

This PR fixes that issue, adding MarshalYAML methods where needed, and switching to the use of (Un)MarshalText where possible. Both YAML and JSON respect the TextMarshaler interfaces; for types which we marshal as their string representation, e.g. StoragePolicy, this saves a bit of boilerplate.

N.B.: I haven't fixed this issue for every type it affects; unfortunately, the fix + testing is pretty manual, despite the helpers. I can file a separate issue for that if desired.

Special notes for your reviewer:

Does this PR introduce a user-facing and/or backwards incompatible change?:
NONE

Does this PR require updating code package or user-facing documentation?:
NONE