[build] Ignore modification time when generatking statik files (#2019)
Running `make services` will eventually execute the `build-ui-ctl-statik` target. This target unconditionally runs `statik` to generate the file `src/ctl/generated/ui/statik/statik.go`. By default, `statik` will include the time at which it is run in the file that it generates. This has the unfortunate consequence that running `make services` modifies the code base though. Since we unconditionally run `statik` we can set the `-m` flag to ensure that `statik` includes a constant modification time in the output that it generates, thereby ensuring that the file it generates is deterministic.