Support custom grouping of instances by mirrored placement algorithm (#2074)
This PR makes the algorithm for grouping hosts into shardsets for the mirrored placement algorithm pluggable, and introduces a simple "group id" based strategy as an alternative The current behavior is:

Group instances by their port--e.g. 5555 with 5555, ensuring that:
a. Instances are on a different host and in a different isolation group
b. Instances have the same weight.

I've left the default behavior the same while introducing a new option, SetInstanceSelector to support changing the behavior.