Add an in memory implementation of cluster/client.Client (#2219)
What this PR does / why we need it:

This adds a small testing utility to back a cluster/client.Client with an in-memory kv.Store. We have the cluster/kv/mem implementation of kv.Store, but there's currently no way to back a cluster/client.Client with it. Construction of the kv.Store is tightly coupled with etcd in the current class; having a separate class seemed like a reasonable way to do this given the current code, but we could consider a more dependency injection style for that--let me know if you want to hash that out.