Use random ports for etcd (#2244)
What this PR does / why we need it:

Currently, our embedded KV module uses etcd's default ports. This is problematic in a CI environment; we can't guarantee openness of those ports. Better is to use a random port, which gives us test isolation.

Note: if there's some code assuming the default ports, this will break it. I'm not concerned about that because:

This is a change to a test module.
Users can already reference the endpoints using Endpoints()
If we don't want the backwards incompatible change though, I can put it behind a flag.
Special notes for your reviewer:

Does this PR introduce a user-facing and/or backwards incompatible change?:

Changes src/cluster/integration to use random ports for etcd. The server addresses can be accessed via `Endpoints()`.
Does this PR require updating code package or user-facing documentation?:

NONE