[m3admin] API: Support namespace retention updates (#2383)
We currently don't have a way to modify a namespaces retention with the
m3admin APIs. However doing so is somewhat tricky, as it's unsafe to
update any arbitrary namespace field.

This PR supports safely updating a namespace in-place. The only
respected field is retention period, as we know that to be safe.
Attempts to modify any other field will inform the user the field is
immutable.