Bump etcd client to 3.4.10, remove GRPC and etcd replace directives (#2555)
What this PR does / why we need it: This PR updates the etcd client version we use to work with later versions of GRPC. It also removes the replace directives for both GRPC and etcd.

My suspicion based on talking with @notbdu is that these replace directives aren't strictly necessary, and were a function of dependency import from glide. Removing them hopefully will make version maintenance easier (no guarantees).

A note about the etcd dep and version: etcd 3.4 is in a weird state wrt go modules; tl;dr is that the switch to go modules for go 3.4 broke etcd as a library with go modules; see etcd-io/etcd#11154 for details. The recommendation from that thread is to use a hash based version string (method described in etcd-io/etcd#11154 (comment)). We were previously doing this in a replace directive; I've switched the dependency to do it in the require block directly. IIUC, this propagates the dependency to downstream code using m3 as a library, whereas the replace directive does not.