Cache the results of the commitlog bootstapper between runs
A recent refactoring of cold writes ( https://github.com/m3db/m3/pull/2508/) introduced a regression
that results in the commit log being read twice while bootstrapping. The referenced PR changed the
commitlog source to read all time ranges, not just the requested time ranges. This was necessary since
the commitlog may have cold writes that were never commmited to a fileset. Another bootstrapper (filesystem or peer)
might report a ShardTimeRange as fulfilled, but it could be missing recent cold writes from a commit log.

The bootstrapper is split into 2 runs, one for the cold time ranges and one for the recent warm time range. Since the commitlog source
now reads the full time range for both runs, it resulted in 2 full reads of the commit log/snapshots.

This commit introduces a simple fix that caches the results of the commitlog source between runs. Other options
were considered, such as adding an option to disable a specific bootstrapper. This change seemed the most ideal since it
isolates the special logic of the commitlog bootstrapper to a single file.