Only read the commit log once during bootstrapping (#2645)
* Only read the commit log once during bootstrapping

A recent refactoring of cold writes ( #2508) introduced a regression that increases the
chances the commit log is read twice while bootstrapping. The referenced PR changed the commitlog
bootstrapper to read all requested time ranges, even if a range had been fulfilled by a previous
bootstrapper. This was necessary since the commitlog may have cold writes that were never commmited
to a fileset. The fileystem bootstrapper would report a time range as fulfilled, but might be missing
cold writes only in the commit log.

It should be noted this bug was always theoretically possible, but unlikely since the commitlog bootstrapper
typically wouldn't run in the first pass (cold time ranges) since the filesystem would fulfill all cold ranges
and short circuit the first pass of the boostrapper.

This change only reads the commit log on the first pass of the boostrapper and caches the result to skip reading
it in subsequent passes. It doesn't actually matter which pass we read the commit log, the first was just chosen
arbitrarily.

Other attempts at fixing this bug attempted to disable the entire commit log bootstrapper during a pass, but that's
not possible since the commit log bootstrapper is actually 2 bootstrappers in one, both the the commit log and snapshot
files. To minimize the refactoring changes we still want to only read the snapshot files of the requested ranges.