Add metrics for bytes added/removed to the m3msg buffer
Helpful for understanding how to size the buffer to survive a m3aggregator
deploy. Due to the graceful shutdown of a m3agregator, it can take almost a
minute between connections closing on the old instances and connections
start taking traffic on the new instances. If a coordinator takes 10MB/s
of traffic, it only takes 10s to fill up a 100MB buffer and start dropping requests.