[deps] grpc v1.27 -> v1.29
We're stuck in a tough spot with our gRPC version. We are on a
relatively old version, and gRPC frequently introduces breaking changes
to APIs that are experimental. Our own query code, and the version of
etcd we depend on, use experimental APIs which were later deprecated.

We can't upgrade too far as we won't be able to import etcd. However
v1.29 is far more compatible with other parts of the ecosystem, and
makes it easier to integrate M3 in other environments (as well as
integrate newer tools into M3).