Add a configurable write timeout for consumer ACKs (#2675)
Without a timeout this can potentially block forever.

The producer and consumer can potentially dead lock trying to send/receive messages:

Producer -> msg -> Consumer. Consumer is not attempting to read since it's trying to ACK.
Consumer -> ack -> Producer. Producer is not attempting to read since it might be stuck trying to get a lock on the connection.

For backwards compatibility this defaults to 0. We should set it to a sane default (5s) in the future.