Only hold the writeState ex lock when resetting
the writeState lock is shared across all goroutines, so we need to
avoid grabbing the exclusive write lock unless it's absolutely necessary.

Before this change the write lock was acquired when:
1. Reading some state. This can just use the shared read lock.
2. Flushing. Now the read lock is used to grab the connections and the individual
connection locks are used to serialize writes on an individual connection. Additionally,
closing now grabs the individual connection locks to prevent closing a connection
being used.