Add implementation for dynamically fetching ClusterNamespaces
This PR implements dynamicCluster to support fetching ClusterNamespaces
from etcd dynamically instead of pulling from static config. Notably,
this does not wire up the implementation for use in the coordinator yet.
That change is pending the addition of a flag to namespace metadata
to notify the coordinator that the namespace is ready for use.