Remove RestartSec from m3dbnode.service example
Having `RestartSec=10s` makes systemd wait for 10s before starting the
service back up after it already exited. This does not appear to be the
original intention of it. As we want m3dbnode started as soon as
possible once it stopped on error.

This was intended to be `TimeoutStopSec` which would have waited for 10s
before forcibly SIGKILLing the process. Without it specified the
default is given by `DefaultTimeoutStopSec` which is usually `90s`.

Remove the RestartSec, which would mean it would only wait 100ms before
starting the process after it exited with non-zero status. Do not add
`TimeoutStopSec` so the kill behavior is not changed. If someone wants
it, they can add that option.