AddToReset config to change Add transforms to Reset transforms (#2817)
This is a temporary option to allow seamlessly rolling out Add -> Reset
transforms. With this we can deploy the new aggregator with the option
without experiencing another degradation in aggregated counters. Without
the change we'd need to deploy the new code and then deploy the new rules.
The deploy would result in degraded aggregated counters without this config.