[lint] Disable nlreturn linter (#2865)
Keeping the new line required before a return linter would require a large fraction of the code base to need updating. 

The perceived benefit of always an empty line does not seem worth the changes required to each file. Also it is purely stylistic and doesn't seem to avoid types of errors.