Consecutive replaces in mirrored placement
Problem:

This fixes issue #2850. The mirrored placement requires all shards to be available to process an instance replace. This makes it impossible to have a consecutive replace following an earlier replace until *all* shards in placement are cutover, even if the scopes of replaces do not overlap, i.e. the replace pairs own disjoint sets of shards. In practice it significantly slows down consecutive replaces and increases the risk of data loss because the longest supported aggregation tile is 1 hour.

Solution:

When processing a replace, require only the leaving instance and its peers to have their shards available. The peer instances are instances that own the same shardset which includes mirror peers (when replication factor >= 2) and any pending replaces where the specified leaving node can be either a replacement or replaced.

Add new tests asserting this use case.