Refactor FetchTagged to return an Iterator of results (#3141)
This is the first step in several refactorings to limit how many series blocks can be loaded at once. This will prevent
large queries from overwhelming the system and give more fair access to all queries in the system.

This first refactoring creates the interface for callers to use to iterate through series blocks one at a
time. The series blocks are still all loaded at once and this will be fixed in a future PR with another Iterator.