Add background cancel monitor for index query
Use a background monitor to cancel the index query after the timeout.
This prevents a race in the previous code that only checked the timeout
when scheduling a new worker. Once a worker was dispatched it could in
theory run forever since it would not recheck the timeout.

Now the background goroutine will cancel the worker after the timeout.
The worker checks the cancelation after each batch, so at worst a worker
will process one more additional batch after the timeout.