[dbnode] Use go context to cancel index query workers after timeout (#3194)
This prevents a race in the previous code that only checked the timeout
when scheduling a new worker. Once a worker was dispatched it could in
theory run forever since it would not recheck the timeout.

Since the thrift handler always includes a go context with a timeout, we can use it to cancel the workers. A lot of tests were not setting the go context in the m3 context, so this includes an additional mechanical refactoring to set a Background go ctx in all tests.