Fix data race in genny list Remove
The data race is reading the value after the element has been put back
in the pool. Another goroutine can get the element out of pool before
the value is read again.

This was not causing an actual production issue because no caller
actually uses the returned value from Remove. However, it did cause
data race failures in tests when run with -race.