Eagerly search all segments for an index block
This is part of a larger refactoring to chunk up the processing of a single
block. Each chunk would need to acquire an index worker from the pool.
This prevents large queries with many documents results in an index
block from starving smaller queries. Ideally every chunk of work would
hold an index worker for the same amount of time.

In this change we eagerly search all the segments for a block on the
first call to Next(). Metrics show the time search is relatively
constant across queries, so it shouldn't add much overhead to the first
Next() call. Once the postings lists are eagerly retrived, the doc
iterator can iterate over them in batches and yield the worker after N
docs. Once it acquires a worker again it can resume the iteration.