Add HasNext to search iterator
This allows peeking if Next() will return true on the next call. The
index query path will use this method in the future to check if it
should wait for an index worker from the pool. Once it acquires the
worker it can start iterating with Next(). It would silly to wait for an
index worker if there is nothing to do.