Make staged placement storage behave as non-staged in backward-compatible way
Problem:

M3 aggregator uses staged placement that maintains a collection of placement snapshots in the same etcd value payload. This does not scale for large deployments of 1K+ aggregator nodes and O(10K) clients:
* Each such snapshot is large by itself due to the large number of nodes and even larger number of shards of aggregator.
* When multiple node replaces happen, the staged placement ends up holding multiple placement snapshots, including those with cutover time before now and those with cutover time > now.
* etcd gets overloaded and is not capable to propagate payload that large to all clients.
* Clients that don't get update from etcd still use stale version that increases the risk of data loss when aggregator nodes are replaced.

Feel free to correct me, but there seem to be no practical use case for staged placements. If there is a clock skew among the clients, having placements activated in the future (via `SetPlacementCutoverNanosFn`) is not going to help all clients use that placement all at the same time and is not different if the placement was to become active at the time of update.

The shards have their own cutover/cutoff time that is orthogonal to placement cutover time. Having single latest placement is sufficient to properly handle movements of shards between the nodes upon node replacements.

Solution:

Change implementation of placement storage to keep only single latest snapshot of the staged placement when it is updated.
This ensures backward compatibility: when existing deployment with multiple snapshots in etcd is upgraded to this version of the code, the next update of the placement will delete all snapshots except the latest and set its cutover time to `0` effectively making it the active placement.