Refactor block.Aggregate to use a doc Iterator (#3256)
This is mostly a refactoring change to prove the Iterator pattern
behaves the same as the previous impl. The one semantic change is the
segments are searched eagerly on the first call to Next() like the Query Iterator.

This change is was involved than block.Query since a wrapper Iterator to
process an entire block did not exist and had to be created.

In a future commit, the Aggregate method will be removed and the index workers
will directly invoke AggregateWithIter to process a chunk of the Iterator at a time.

* Remove fieldAndTermsIterator.Reset

This is no longer necessary because we don't reuse the iterator between
segments. Instead we search all segments eagerly, so need an iterator
per segment.