Apply NaN filtering while rendering series (#3297)
* Fix duplicate results bug due to non-idempotent render function

The render result function would filter out NaN datapoints before
iterating them to render. This mutated the underlying series array
and made this function unsafe to call more than once. This was not
known when the function was called twice in a row to support
returned-data limits. The flow is to call once to check limits
and once to actually write results (in order to write the limit
header before the response body). This resulted in the series list
getting clobbered on the subsequent render call since it had been
mutated by the first, leading to duplicate results. The duplicates
were also non-deterministic in appearing because the received series
list itself is unordered and so the series with NaN were not in the
same positions on each query.

The solution here is to avoid this filtering step that mutates the
series and itself check for NaN as the series are iterated for
rendering.