Restrict the time a query can hold an index worker (#3269)
A new MaxWorkerTime option is introduced to cap how long a query can hold an 
index worker at once. If the max is exceeded, the
query must yield the index worker back and acquire it again
(potentially waiting) to continue processing the results.

This cap ensures large queries don't dominate the finite number of index
workers allowed to run concurrently and lock out smaller queries. The
idea is users would want to set the max large enough so the vast
majority of typical queries can finish with only a single worker
acquisition.