Add filesystem reporter for available space
Currently, there are no metrics reporting on the
amount of space available in the filesystem directory
path used by the dbnode. Add a reporter which
reports on the available space. This metric could be
used for alerting purposes if the available space
crosses thresholds.

Example metrics reported:

filesystem_avail_bytes 4.3406266368e+10
filesystem_num_api_errors 0
filesystem_total_bytes 6.2725623808e+10

Which matches with the output of df
1K-blocks      Used Available
61255492  15725236  42388932