Do not use buffer channels if growOnDemand is true
This ensures a new worker is spawned if all workers are currently busy.
With buffered channels, a unit of work might get scheduled on a channel
where the worker is currently processing a long request. That unit of
work will get starved until the long request is complete.

For example in the host_queue worker pool, a long query request might
take 2 mins to complete. An unrelated request might get buffered on that
same chhanel and won't get processed for 2 mins.