Wrap errors from the m3 remote storage (#3427)
Currently when the prometheus query execution returns an error, it is
propagated as is. If the error happened in the prometheus library, it
will be returned as a 500, since m3 does not know how to handle
prometheus errors. If the error happened in the prom library, and not in
the m3 storage layer, it is most likely due to a bad query and should be
returned as a 4xx.

This change wraps all errors that come back from the m3 storage layer,
so the callers of the prom library (i.e http handlers) can determine if
the error happened in prom or in m3. If the error happened in prom, it
is returned as a 4xx to the user.