Allow rollup rules with empty group bys (#3458)
An empty groupBy is useful if a user wants a rollup multiple metrics
into a single metric.

For example http_requests and http_errors into something like http_all.

This required fixing a bug in the rollupIDProvider which assumed the
tagPairs was non-empty. Now the nameTagIndex and rollupTagIndex start
at 0, instead of -1, to handle the empty tagPairs case.