Add InstanceMultiple config to lower per instance series limit (#3516)
Previously the per query series limit was passed directly to every
database instance as the series limit.
For large clusters this can result in a lot of wasted series read from the
database. For example a clsuter with 30 instances per replica could read 30x
the time series, just to truncate the results in the coordinators.

Theoretically if the data is perfectly sharded, the per instance limit
is equal to the query series limit / # instance per replica. To allow
for unbalanced data, the InstanceMultiple config increases this
theoretical calculation by a factor to add a buffer.

By default, this option is off, but the hope is to add a sane default
for the next major release.

Additionally if this config is set, a bug is fixed that never returned
an error if RequiredExhaustive=true and the query series limit was
exceeded by summing the individual instance results in the coordinator.
This bug, coupled with the larger per instance series limt, would
actually result in the coordinator processing far more results than
necessary and then silently truncating the results.